backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pyodbc
opencv-python-headless
ultralytics
passlib[bcrypt]
python-jose[cryptography]
python-multipart
azure-storage-blob
pandas
pydantic[email]
pydantic-settings
twilio
requests
scikit-learn
-e ./Yolov5_StrongSORT_OSNet
onnxruntime==1.20.0
onnx